// apps/client/src/types/payment.types.ts
export interface PaymentStatsApiResponse {
  totalRevenue: number
  paymentMethods: {
    stripe: {
      count: number
      amount: number
      percentage: number
    }
    cash?: {
      count: number
      amount: number
      percentage: number
    }
    "mobile-money"?: {
      count: number
      amount: number
      percentage: number
    }
  }
  statuses: {
    failed: {
      count: number
      amount: number
    }
    refunded: {
      count: number
      amount: number
    }
    pending: {
      count: number
      amount: number
    }
    completed: {
      count: number
      amount: number
    }
  }
  dailyRevenue: {
    _id: string
    total: number
    count: number
  }[]
  averagePaymentAmount: number
  refundStats: {
    totalRefunded: number
    count: number
    averageRefund: number | null
  }
}